+++ tests/testthat/test-offset.R
@@ -160,6 +160,22 @@ test_that("offset gets passed through cross validation as expected #372", {
   expect_true(length(unique(y$cv_predicted)) == 684L)
 })
 
+test_that("predicting on newdata with a non-null offset in fit but a null offset in predict informs the user appropriately", {
+  dat <- subset(dogfish, catch_weight > 0)
+  fit <- sdmTMB(
+    catch_weight ~ 1,
+    data = dat,
+    family = Gamma("log"),
+    offset = "area_swept",
+    spatial = "off"
+  )
+  pred <- predict(fit)
+  pred <- predict(fit, offset = rep(0, nrow(dat)))
+  pred <- predict(fit, newdata = qcs_grid, offset = rep(0, nrow(qcs_grid)))
+  pred <- predict(fit, newdata = qcs_grid)
+  expect_message({pred <- predict(fit, newdata = qcs_grid)}, regexp = "offset")
+})
+
 # #
 # # offset/prediction setting checks:
 #